+++ tests/test_lognflow.py
@@ -268,9 +268,9 @@ def test_log_single_text():
     logger = lognflow(temp_dir)
     logger('This is a test for test_log_single_text', flush = True)
     var = 2
-    logger.log_single('text_log', 'hello\n', save_as='txt', time_tag = False)
-    logger.log_single('text_log', 'bye\n', save_as='txt', time_tag = False)
-    logger.log_single('text_log', var, save_as='txt', time_tag = False)
+    logger.log_single('text_log\a\t/\b/\b//\\/b', 'hello\n', save_as='txt', time_tag = False)
+    logger.log_single('text_log\a', 'bye\n', save_as='json', time_tag = False)
+    logger.log_single('text_log\a', var, save_as='pdb', time_tag = False)
     
 def test_log_imshow_complex():
     logger = lognflow(temp_dir)
@@ -287,10 +287,10 @@ if __name__ == '__main__':
     #-----IF RUN BY PYTHON------#
     temp_dir = select_directory()
     #---------------------------#
+    test_log_single_text()
     test_log_imshow_complex()
     test_log_imshow()
     test_log_text()
-    test_log_single_text()
     test_log_surface()
     test_log_single()
     test_lognflow_conflict_in_names()
@@ -308,4 +308,3 @@ if __name__ == '__main__':
     test_log_hexbin()
     test_log_canvas()
     test_log_confusion_matrix()
-    exit()